Render registration form as div to avoid accessibility error
3.2.2 On Input
Description of this issue:
The page contains a form but there is no submit button.
How to fix it:
The form should be provided with a submit button to maximize predictability.
Our solution: Render a div instead of a form element, since all
the buttons are links rather than inputs.