Remove the toolbar access from course admins
The content overview page is a bit confusing, b/c the 
SalesForce module fails to disable the operation for nodes
that the course admin can't edit. So there are a bunch of
mapped nodes with a "Salesforce" operation listed. Clicking
on it returns "Access Denied". This role can do so little
that the toolbar isn't really needed.