Add class session inline entity form to class node forms
To do so, I had to set default form display options on some base fields.

At the moment, the daterange field doesn't work properly because the values imported from salesforce are missing the times. I know that field might change in sf, so I'll leave it for now.

Lastly, the inline entity forms are not disabled. I suspect it's an issue with the salesforce_mop module.