Transform embedded file entities into embedded `document` media entities during import
This reverts most of e913c72. We're going back to media entities because users must be able to select them from a media library.

Note that this reversion actually adds a bit of functionality in that it includes the original link text in a new `data-link-text` attribute. With luck, we'll be able to use this value to customize the embed media modal like images do with their alt text field. The idea will be to add a 'Link text' value. Our ability to do this remains to be seen.