Update basic formatting with media text format config after export
After our update of core from 9.2.6 to 9.2.10, saving and exporting this config re-orders the allowed elements and modifies the allowed classes.

In this case `class` was changed to `class=""` in some cases, and items in the stylescombo plugin (AKA 'Styles dropdown') had their classes added automatically. For example, the <a> element was set to `class=" cu-button"`.

We don't want that. We want to continue to allow all classes on the <a> element. So we removed the cu-button value and the filter appears to work as it did previously.