Move the persistent_visitor_parameters options to a prod-only services file
This should allow the pvp cookie to be properly set on non-production sites. When the cookie_domain setting is NULL, as it is by default in persistent_visitor_parameters.services.yml, the current hostname should be set for the cookie domain.

But for production, we should continue to use the .ilr.cornell.edu domain, so that cookies will be sent to other ILR web sites.