Add the custom utm_* query param webform script to the admin-displayed snippet
This is not required. It's only here for consistency with the marketing theme. It might also help when cutting and pasting the code manually, although the script.js embed from the marketing site should automatically include this template from the marketing theme and this this js code.

See commit cb601e2. This could probably be symlinked to the template from that commit to reduce confusion.