WIP: Add example of replacing a Section Heading with a Section in a custom block template
This is tricky, because a lot of these are place in containers that are custom layouts that use `cu-section` or similar.

We may need to keep the Section Heading component to allow placement after other items in a section. Or perhaps the design of some of these pages, especially posts will change.