Allow in-page nav fragments on all sections
This adds a new `fragment` setting in addition to the `title` setting.

Previously, `title`, if set, would be automatically converted into a fragment.

But the title option was hidden if the host entity didn't have the section navigation extra field in its display. Therefore, entities that did not display the in-page nav also couldn't set ids on the section elements, and had to resort to rich text field hacks to enable in-page anchors.

Now, the fragment option is always available, but the title option is only visible if the host entity supports the in-page nav extra field.