Fix session date timezone issue on course pages
The twig `date` function uses the PHP default timezone, which is generally America/New_York on our systems. Converting a date from Drupal storage without explicitly setting the timezone will assume that it's the default, rather than UTC, which is Drupal's storage timezone.

So class dates late in the day would appear to be the next day. Confoundingly, since the time was displayed using the `drupal_field` function, the timezone was correct and the time displayed accurately.