Allow standalone access to media entities
This was done to allow \Drupal\ilr\PathProcessor access to rewrite canonical media links, and fix a bug with rewriting paths for remote_file media entities.

When `standalone_url` is false, the Media canonical link template (and thus route) is the same as the edit link template and route. This makes it impossible to tell if the route is for viewing or editing the Media entity.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>