Allow form admins to see the results of the attending online checkbox
If #access is set to `false` for an element, it is excluded from the submission results. The downside of this approach is that webform does not correctly interpret the checkbox value, so submissions show a 1 or 0. However, that's a better alternative than having to go item by item to see the results.