Add deprecation note for the disable_salesforce_submitter variant
We should probably remove this, but it would throw a notice on every webform that has it in the config. Currently, that's quite a few. 

Additionally, the presence of this handler might be confusing, since it is actually the eventid that determines whether it is sent as a touchpoint. We could configure it in a way that prevents the touchpoint from being sent, even though there's no compelling reason to have two different ways of preventing it from being sent to SF.