WIP: Add 'modal' variable to persona templates
The idea is to use this to load a different template for default personas when loaded in a modal, e.g. use the profile template for a direct link but the modal (or full, if we remove modal) template within the modal.

This should work because modal requests include the `_wrapper_format` query parameter and should be therefore be cached separately.