Add 'modal' variable to persona templates
This reverts most of the persona modal view mode code in favor of using a different template for default personas when loaded in a modal, e.g. use the profile template for a direct link but the full template within the modal.

This works because modal requests include the `_wrapper_format` query parameter and should be therefore be cached separately.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>