Remove requirement for image file field on logo media entities
While it may seem strange to remove the requirement for the actual logo image file, these logos are mainly used on media mentions, which no longer use the logo image. They use the logo `name` field instead. This spares content editors from needing to hunt down a logo that they know won't be used.

We decided to keep this as a media entity because we may, in the future, add a field indicating that we have permission to use the logo, in which case we may return to displaying it.