Prevent duplicate person entities during recurring ilr_employee_personas migration
Duplicates were being created because the migration would reset the `person` base field on ilr_employee personas since the static_map defaults to null.

This would trigger code in Persona::preSave() that would create a new person when the `person` base field had no value.

To prevent the overriding of the `person` base field, we explicitly list the fields to be overwritten in the destination settings, and we ensure that the `person` property is not included.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>