Fix resource library filter facets
This was caused by the removal of the FacetsArea views plugin between facets 3 beta1 and beta3.

We skipped using the official facet views filters and continued to use the provided blocks, simply moving them from the view header to the resource library collection default layout template.

Note that this makes it very hard to edit the layout because of UI bugginess.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>