Require Organization field on NCRS webform
Although the Touchpoint does not require this field, at least one Salesforce Flow copies the Organization/Company field to a Lead, where that field is required. When that happens, the Flow fails if this value is empty.

Alternatively, we could set a default value when empty (they prefer 'NONE PROVIDED' but we would need to dispatch an event or add a hook to TouchPointHandler to accommodate that.