Remove honeypot protection from Webforms
Honeypot triggers the page_cache killswitch on any page with a protected form. Unfortunately for us, most subsites have a contact form on every single page.

See HoneypotService::addFormProtection().

The plan is to try to add a captcha, which most likely will kill the cache when added to forms. We'll see.

Note that some Webforms are not exported to config/sync. That's why there's a deploy hook to catch those.