WIP: Update post listing paragraph for use in Paragraphs Browser
This could instead be a commit where we add decent descriptions to all of the paragraphs that can appear in Sections.

Note that the image path for the paragraphs browser must be a file in the web directory. We may need to take additional steps to serve these items with proper caching and file size settings.