Add deploy hook to delete the milr_visit_form_submission_lead salesforce mapping
Even though the mapping was removed from config, our config ignore pattern module ignores the removal and leaves the mapping behind in the database.

This also removes the mapped objects that would otherwise have been orphaned.

It retains the 'mapped' salesforce object reference in the key_value store, which shows up in the notes of the previous submissions.